Created a basic messaging interface based on GiftedChat
A major part of this was getting the keyboard to not cover the text
input on Android. GiftedChat proposed a simple solution using
KeyboardAvoidingView, however this did not work as documented and the
text input was still obscured. In the end, the solution was found in
the blog post linked at the bottom of the commit message. This
solution creates a KeyboardShift component, which keeps track of the
keyboard opening and closing and scrolls the view accordingly. This
solution acts slightly unlike a user would expect in that it scrolls
the entire view including the title bar, however this seems like the
lesser of two evils.

The true evils here are of course this projects' 800MB `node_modules/`
directory and javascript, may they return the the deepest darkest
parts of hell from which they came.

Aforementioned blog post:
https://codeburst.io/react-native-keyboard-covering-inputs-72a9d3072689